requirements.txt
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
xlrd==1.2.0
